@@ -1,10 +1,10 @@
 {
     "name": "bacon/bacon-qr-code",
     "description": "BaconQrCode is a QR code generator for PHP.",
-    "license" : "BSD-2-Clause",
+    "license": "BSD-2-Clause",
     "homepage": "https://github.com/Bacon/BaconQrCode",
     "require": {
-        "php": "^7.1 || ^8.0",
+        "php": "^8.1",
         "ext-iconv": "*",
         "dasprid/enum": "^1.0.3"
     },
@@ -24,15 +24,21 @@
             "BaconQrCode\\": "src/"
         }
     },
+    "autoload-dev": {
+        "psr-4": {
+            "BaconQrCodeTest\\": "test/"
+        }
+    },
     "require-dev": {
-        "phpunit/phpunit": "^7 | ^8 | ^9",
-        "spatie/phpunit-snapshot-assertions": "^4.2.9",
-        "squizlabs/php_codesniffer": "^3.4",
-        "phly/keep-a-changelog": "^2.1"
+        "phpunit/phpunit": "^10.5.11 || 11.0.4",
+        "spatie/phpunit-snapshot-assertions": "^5.1.5",
+        "squizlabs/php_codesniffer": "^3.9",
+        "phly/keep-a-changelog": "^2.12"
     },
     "config": {
         "allow-plugins": {
-            "ocramius/package-versions": true
+            "ocramius/package-versions": true,
+            "php-http/discovery": true
         }
     },
     "archive": {
