@@ -16,21 +16,13 @@ final class BitArray
      *
      * @var SplFixedArray<int>
      */
-    private $bits;
-
-    /**
-     * Size of the bit array in bits.
-     *
-     * @var int
-     */
-    private $size;
+    private SplFixedArray $bits;
 
     /**
      * Creates a new bit array with a given size.
      */
-    public function __construct(int $size = 0)
+    public function __construct(private int $size = 0)
     {
-        $this->size = $size;
         $this->bits = SplFixedArray::fromArray(array_fill(0, ($this->size + 31) >> 3, 0));
     }
 
@@ -107,7 +99,7 @@ final class BitArray
         }
 
         $result = ($bitsOffset << 5) + BitUtils::numberOfTrailingZeros($currentBits);
-        return $result > $this->size ? $this->size : $result;
+        return min($result, $this->size);
     }
 
     /**
@@ -133,7 +125,7 @@ final class BitArray
         }
 
         $result = ($bitsOffset << 5) + BitUtils::numberOfTrailingZeros($currentBits);
-        return $result > $this->size ? $this->size : $result;
+        return min($result, $this->size);
     }
 
     /**
