@@ -11,24 +11,8 @@ namespace BaconQrCode\Common;
  */
 final class EcBlock
 {
-    /**
-     * How many times the block is used.
-     *
-     * @var int
-     */
-    private $count;
-
-    /**
-     * Number of data codewords.
-     *
-     * @var int
-     */
-    private $dataCodewords;
-
-    public function __construct(int $count, int $dataCodewords)
+    public function __construct(private readonly int $count, private readonly int $dataCodewords)
     {
-        $this->count = $count;
-        $this->dataCodewords = $dataCodewords;
     }
 
     /**
