@@ -13,22 +13,14 @@ namespace BaconQrCode\Common;
 final class EcBlocks
 {
     /**
-     * Number of EC codewords per block.
-     *
-     * @var int
-     */
-    private $ecCodewordsPerBlock;
-
-    /**
      * List of EC blocks.
      *
      * @var EcBlock[]
      */
-    private $ecBlocks;
+    private array $ecBlocks;
 
-    public function __construct(int $ecCodewordsPerBlock, EcBlock ...$ecBlocks)
+    public function __construct(private readonly int $ecCodewordsPerBlock, EcBlock ...$ecBlocks)
     {
-        $this->ecCodewordsPerBlock = $ecCodewordsPerBlock;
         $this->ecBlocks = $ecBlocks;
     }
 
