@@ -21,14 +21,8 @@ final class ErrorCorrectionLevel extends AbstractEnum
     protected const Q = [0x03];
     protected const H = [0x02];
 
-    /**
-     * @var int
-     */
-    private $bits;
-
-    protected function __construct(int $bits)
+    protected function __construct(private readonly int $bits)
     {
-        $this->bits = $bits;
     }
 
     /**
