@@ -10,20 +10,8 @@ use BaconQrCode\Renderer\Path\Path;
  */
 final class CompositeEye implements EyeInterface
 {
-    /**
-     * @var EyeInterface
-     */
-    private $externalEye;
-
-    /**
-     * @var EyeInterface
-     */
-    private $internalEye;
-
-    public function __construct(EyeInterface $externalEye, EyeInterface $internalEye)
+    public function __construct(private readonly EyeInterface $externalEye, private readonly EyeInterface $internalEye)
     {
-        $this->externalEye = $externalEye;
-        $this->internalEye = $internalEye;
     }
 
     public function getExternalPath() : Path
