@@ -16,18 +16,11 @@ final class DotsModule implements ModuleInterface
     public const MEDIUM = .8;
     public const SMALL = .6;
 
-    /**
-     * @var float
-     */
-    private $size;
-
-    public function __construct(float $size)
+    public function __construct(private readonly float $size)
     {
         if ($size <= 0 || $size > 1) {
             throw new InvalidArgumentException('Size must between 0 (exclusive) and 1 (inclusive)');
         }
-
-        $this->size = $size;
     }
 
     public function createPath(ByteMatrix $matrix) : Path
