@@ -5,20 +5,8 @@ namespace BaconQrCode\Renderer\Path;
 
 final class Move implements OperationInterface
 {
-    /**
-     * @var float
-     */
-    private $x;
-
-    /**
-     * @var float
-     */
-    private $y;
-
-    public function __construct(float $x, float $y)
+    public function __construct(private readonly float $x, private readonly float $y)
     {
-        $this->x = $x;
-        $this->y = $y;
     }
 
     public function getX() : float
@@ -37,5 +25,18 @@ final class Move implements OperationInterface
     public function translate(float $x, float $y) : OperationInterface
     {
         return new self($this->x + $x, $this->y + $y);
+    }
+
+    /**
+     * @return self
+     */
+    public function rotate(int $degrees) : OperationInterface
+    {
+        $radians = deg2rad($degrees);
+        $sin = sin($radians);
+        $cos = cos($radians);
+        $xr = $this->x * $cos - $this->y * $sin;
+        $yr = $this->x * $sin + $this->y * $cos;
+        return new self($xr, $yr);
     }
 }
