@@ -28,14 +28,8 @@ final class PlainTextRenderer implements RendererInterface
      */
     private const EMPTY_BLOCK = "\xc2\xa0";
 
-    /**
-     * @var int
-     */
-    private $margin;
-
-    public function __construct(int $margin = 2)
+    public function __construct(private readonly int $margin = 2)
     {
-        $this->margin = $margin;
     }
 
     /**
