@@ -7,26 +7,11 @@ use BaconQrCode\Renderer\Color\ColorInterface;
 
 final class Gradient
 {
-    /**
-     * @var ColorInterface
-     */
-    private $startColor;
-
-    /**
-     * @var ColorInterface
-     */
-    private $endColor;
-
-    /**
-     * @var GradientType
-     */
-    private $type;
-
-    public function __construct(ColorInterface $startColor, ColorInterface $endColor, GradientType $type)
-    {
-        $this->startColor = $startColor;
-        $this->endColor = $endColor;
-        $this->type = $type;
+    public function __construct(
+        private readonly ColorInterface $startColor,
+        private readonly ColorInterface $endColor,
+        private readonly GradientType   $type
+    ) {
     }
 
     public function getStartColor() : ColorInterface
