@@ -10,40 +10,19 @@ use BaconQrCode\Renderer\Module\SquareModule;
 
 final class RendererStyle
 {
-    /**
-     * @var int
-     */
-    private $size;
+    private ModuleInterface $module;
 
-    /**
-     * @var int
-     */
-    private $margin;
+    private EyeInterface|null $eye;
 
-    /**
-     * @var ModuleInterface
-     */
-    private $module;
-
-    /**
-     * @var EyeInterface|null
-     */
-    private $eye;
-
-    /**
-     * @var Fill
-     */
-    private $fill;
+    private Fill $fill;
 
     public function __construct(
-        int $size,
-        int $margin = 4,
+        private int $size,
+        private int $margin = 4,
         ?ModuleInterface $module = null,
         ?EyeInterface $eye = null,
         ?Fill $fill = null
     ) {
-        $this->margin = $margin;
-        $this->size = $size;
         $this->module = $module ?: SquareModule::instance();
         $this->eye = $eye ?: new ModuleEye($this->module);
         $this->fill = $fill ?: Fill::default();
