@@ -11,29 +11,15 @@ use SplFixedArray;
 final class BlockPair
 {
     /**
-     * Data bytes in the block.
-     *
-     * @var SplFixedArray<int>
-     */
-    private $dataBytes;
-
-    /**
-     * Error correction bytes in the block.
-     *
-     * @var SplFixedArray<int>
-     */
-    private $errorCorrectionBytes;
-
-    /**
      * Creates a new block pair.
      *
-     * @param SplFixedArray<int> $data
-     * @param SplFixedArray<int> $errorCorrection
+     * @param SplFixedArray<int> $dataBytes Data bytes in the block.
+     * @param SplFixedArray<int> $errorCorrectionBytes Error correction bytes in the block.
      */
-    public function __construct(SplFixedArray $data, SplFixedArray $errorCorrection)
-    {
-        $this->dataBytes = $data;
-        $this->errorCorrectionBytes = $errorCorrection;
+    public function __construct(
+        private readonly SplFixedArray $dataBytes,
+        private readonly SplFixedArray $errorCorrectionBytes
+    ) {
     }
 
     /**
