@@ -14,10 +14,10 @@ final class MaskUtil
     /**#@+
      * Penalty weights from section 6.8.2.1
      */
-    const N1 = 3;
-    const N2 = 3;
-    const N3 = 40;
-    const N4 = 10;
+    public const N1 = 3;
+    public const N2 = 3;
+    public const N3 = 40;
+    public const N4 = 10;
     /**#@-*/
 
     private function __construct()
