@@ -8,26 +8,13 @@ use BaconQrCode\Exception;
 final class Alpha implements ColorInterface
 {
     /**
-     * @var int
-     */
-    private $alpha;
-
-    /**
-     * @var ColorInterface
-     */
-    private $baseColor;
-
-    /**
      * @param int $alpha the alpha value, 0 to 100
      */
-    public function __construct(int $alpha, ColorInterface $baseColor)
+    public function __construct(private readonly int $alpha, private readonly ColorInterface $baseColor)
     {
         if ($alpha < 0 || $alpha > 100) {
             throw new Exception\InvalidArgumentException('Alpha must be between 0 and 100');
         }
-
-        $this->alpha = $alpha;
-        $this->baseColor = $baseColor;
     }
 
     public function getAlpha() : int
