@@ -8,26 +8,11 @@ use BaconQrCode\Exception;
 final class Rgb implements ColorInterface
 {
     /**
-     * @var int
-     */
-    private $red;
-
-    /**
-     * @var int
-     */
-    private $green;
-
-    /**
-     * @var int
-     */
-    private $blue;
-
-    /**
      * @param int $red the red amount of the color, 0 to 255
      * @param int $green the green amount of the color, 0 to 255
      * @param int $blue the blue amount of the color, 0 to 255
      */
-    public function __construct(int $red, int $green, int $blue)
+    public function __construct(private readonly int $red, private readonly int $green, private readonly int $blue)
     {
         if ($red < 0 || $red > 255) {
             throw new Exception\InvalidArgumentException('Red must be between 0 and 255');
@@ -40,10 +25,6 @@ final class Rgb implements ColorInterface
         if ($blue < 0 || $blue > 255) {
             throw new Exception\InvalidArgumentException('Blue must be between 0 and 255');
         }
-
-        $this->red = $red;
-        $this->green = $green;
-        $this->blue = $blue;
     }
 
     public function getRed() : int
@@ -72,6 +53,10 @@ final class Rgb implements ColorInterface
         $m = 1 - ($this->green / 255);
         $y = 1 - ($this->blue / 255);
         $k = min($c, $m, $y);
+
+        if ($k === 0) {
+            return new Cmyk(0, 0, 0, 0);
+        }
 
         return new Cmyk(
             (int) (100 * ($c - $k) / (1 - $k)),
