@@ -15,18 +15,10 @@ use BaconQrCode\Renderer\RendererInterface;
 final class Writer
 {
     /**
-     * Renderer instance.
-     *
-     * @var RendererInterface
-     */
-    private $renderer;
-
-    /**
      * Creates a new writer with a specific renderer.
      */
-    public function __construct(RendererInterface $renderer)
+    public function __construct(private readonly RendererInterface $renderer)
     {
-        $this->renderer = $renderer;
     }
 
     /**
@@ -39,7 +31,7 @@ final class Writer
      */
     public function writeString(
         string $content,
-        string $encoding = Encoder::DEFAULT_BYTE_MODE_ECODING,
+        string $encoding = Encoder::DEFAULT_BYTE_MODE_ENCODING,
         ?ErrorCorrectionLevel $ecLevel = null,
         ?Version $forcedVersion = null
     ) : string {
@@ -62,7 +54,7 @@ final class Writer
     public function writeFile(
         string $content,
         string $filename,
-        string $encoding = Encoder::DEFAULT_BYTE_MODE_ECODING,
+        string $encoding = Encoder::DEFAULT_BYTE_MODE_ENCODING,
         ?ErrorCorrectionLevel $ecLevel = null,
         ?Version $forcedVersion = null
     ) : void {
