@@ -6,17 +6,12 @@ namespace Endroid\QrCode\Color;
 
 final class Color implements ColorInterface
 {
-    private int $red;
-    private int $green;
-    private int $blue;
-    private int $alpha;
-
-    public function __construct(int $red, int $green, int $blue, int $alpha = 0)
-    {
-        $this->red = $red;
-        $this->green = $green;
-        $this->blue = $blue;
-        $this->alpha = $alpha;
+    public function __construct(
+        private readonly int $red,
+        private readonly int $green,
+        private readonly int $blue,
+        private readonly int $alpha = 0
+    ) {
     }
 
     public function getRed(): int
