@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Endroid\QrCode\Encoding;
 
-interface EncodingInterface
+interface EncodingInterface extends \Stringable
 {
     public function __toString(): string;
 }
