@@ -6,18 +6,14 @@ namespace Endroid\QrCode\Label\Font;
 
 final class Font implements FontInterface
 {
-    private string $path;
-    private int $size;
-
-    public function __construct(string $path, int $size = 16)
-    {
-        $this->validatePath($path);
-
-        $this->path = $path;
-        $this->size = $size;
+    public function __construct(
+        private readonly string $path,
+        private readonly int $size = 16
+    ) {
+        $this->assertValidPath($path);
     }
 
-    private function validatePath(string $path): void
+    private function assertValidPath(string $path): void
     {
         if (!file_exists($path)) {
             throw new \Exception(sprintf('Invalid font path "%s"', $path));
