@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Endroid\QrCode\Label;
 
 use Endroid\QrCode\Color\ColorInterface;
-use Endroid\QrCode\Label\Alignment\LabelAlignmentInterface;
 use Endroid\QrCode\Label\Font\FontInterface;
 use Endroid\QrCode\Label\Margin\MarginInterface;
 
@@ -15,7 +14,7 @@ interface LabelInterface
 
     public function getFont(): FontInterface;
 
-    public function getAlignment(): LabelAlignmentInterface;
+    public function getAlignment(): LabelAlignment;
 
     public function getMargin(): MarginInterface;
 
