@@ -6,17 +6,12 @@ namespace Endroid\QrCode\Label\Margin;
 
 final class Margin implements MarginInterface
 {
-    private int $top;
-    private int $right;
-    private int $bottom;
-    private int $left;
-
-    public function __construct(int $top, int $right, int $bottom, int $left)
-    {
-        $this->top = $top;
-        $this->right = $right;
-        $this->bottom = $bottom;
-        $this->left = $left;
+    public function __construct(
+        private readonly int $top,
+        private readonly int $right,
+        private readonly int $bottom,
+        private readonly int $left
+    ) {
     }
 
     public function getTop(): int
