@@ -6,8 +6,6 @@ namespace Endroid\QrCode;
 
 use Endroid\QrCode\Color\ColorInterface;
 use Endroid\QrCode\Encoding\EncodingInterface;
-use Endroid\QrCode\ErrorCorrectionLevel\ErrorCorrectionLevelInterface;
-use Endroid\QrCode\RoundBlockSizeMode\RoundBlockSizeModeInterface;
 
 interface QrCodeInterface
 {
@@ -15,13 +13,13 @@ interface QrCodeInterface
 
     public function getEncoding(): EncodingInterface;
 
-    public function getErrorCorrectionLevel(): ErrorCorrectionLevelInterface;
+    public function getErrorCorrectionLevel(): ErrorCorrectionLevel;
 
     public function getSize(): int;
 
     public function getMargin(): int;
 
-    public function getRoundBlockSizeMode(): RoundBlockSizeModeInterface;
+    public function getRoundBlockSizeMode(): RoundBlockSizeMode;
 
     public function getForegroundColor(): ColorInterface;
 
