@@ -11,14 +11,8 @@ use Endroid\QrCode\QrCodeInterface;
 use Endroid\QrCode\Writer\Result\ConsoleResult;
 use Endroid\QrCode\Writer\Result\ResultInterface;
 
-/**
- * Writer of QR Code for CLI.
- */
-class ConsoleWriter implements WriterInterface
+final class ConsoleWriter implements WriterInterface
 {
-    /**
-     * {@inheritDoc}
-     */
     public function write(QrCodeInterface $qrCode, LogoInterface $logo = null, LabelInterface $label = null, $options = []): ResultInterface
     {
         $matrixFactory = new MatrixFactory();
