@@ -11,29 +11,17 @@ use Endroid\QrCode\QrCodeInterface;
 
 final class DebugResult extends AbstractResult
 {
-    private QrCodeInterface $qrCode;
-    private ?LogoInterface $logo;
-    private ?LabelInterface $label;
-
-    /** @var array<mixed> */
-    private array $options;
-
     private bool $validateResult = false;
 
-    /** @param array<mixed> $options */
     public function __construct(
         MatrixInterface $matrix,
-        QrCodeInterface $qrCode,
-        LogoInterface $logo = null,
-        LabelInterface $label = null,
-        array $options = []
+        private readonly QrCodeInterface $qrCode,
+        private readonly LogoInterface|null $logo = null,
+        private readonly LabelInterface|null $label = null,
+        /** @var array<string, mixed> $options */
+        private readonly array $options = []
     ) {
         parent::__construct($matrix);
-
-        $this->qrCode = $qrCode;
-        $this->logo = $logo;
-        $this->label = $label;
-        $this->options = $options;
     }
 
     public function setValidateResult(bool $validateResult): void
@@ -62,6 +50,7 @@ final class DebugResult extends AbstractResult
             $debugLines[] = 'Logo path: '.$this->logo->getPath();
             $debugLines[] = 'Logo resize to width: '.$this->logo->getResizeToWidth();
             $debugLines[] = 'Logo resize to height: '.$this->logo->getResizeToHeight();
+            $debugLines[] = 'Logo punchout background: '.($this->logo->getPunchoutBackground() ? 'true' : 'false');
         }
 
         if (isset($this->label)) {
