@@ -8,22 +8,22 @@ use Endroid\QrCode\Label\LabelInterface;
 use Endroid\QrCode\Logo\LogoInterface;
 use Endroid\QrCode\QrCodeInterface;
 use Endroid\QrCode\Writer\Result\GdResult;
-use Endroid\QrCode\Writer\Result\PngResult;
 use Endroid\QrCode\Writer\Result\ResultInterface;
+use Endroid\QrCode\Writer\Result\WebPResult;
 
-final class PngWriter extends AbstractGdWriter
+final class WebPWriter extends AbstractGdWriter
 {
-    public const WRITER_OPTION_COMPRESSION_LEVEL = 'compression_level';
+    public const WRITER_OPTION_QUALITY = 'quality';
 
     public function write(QrCodeInterface $qrCode, LogoInterface $logo = null, LabelInterface $label = null, array $options = []): ResultInterface
     {
-        if (!isset($options[self::WRITER_OPTION_COMPRESSION_LEVEL])) {
-            $options[self::WRITER_OPTION_COMPRESSION_LEVEL] = -1;
+        if (!isset($options[self::WRITER_OPTION_QUALITY])) {
+            $options[self::WRITER_OPTION_QUALITY] = -1;
         }
 
         /** @var GdResult $gdResult */
         $gdResult = parent::write($qrCode, $logo, $label, $options);
 
-        return new PngResult($gdResult->getMatrix(), $gdResult->getImage(), $options[self::WRITER_OPTION_COMPRESSION_LEVEL]);
+        return new WebPResult($gdResult->getMatrix(), $gdResult->getImage(), $options[self::WRITER_OPTION_QUALITY]);
     }
 }
