@@ -11,6 +11,6 @@ use Endroid\QrCode\Writer\Result\ResultInterface;
 
 interface WriterInterface
 {
-    /** @param array<mixed> $options */
+    /** @param array<string, mixed> $options */
     public function write(QrCodeInterface $qrCode, LogoInterface $logo = null, LabelInterface $label = null, array $options = []): ResultInterface;
 }
